Backend/requirements.txt
requests
pandas
scikit-learn
torch
fastapi[standard]
uvicorn[standard]
matplotlib
numpy
tqdm
python-dotenv
pydantic